Update route definitions
