+++ api/v1/router.go
@@ -6,9 +6,9 @@ import (
 )
 
 func RegisterRoutes(router *gin.Engine) {
-	router.POST("/create-repo/:token", controllers.CreateRepo)
-	router.DELETE("/delete-repo/:token", controllers.DeleteRepo)
-	router.GET("/list-repos/:username/:token", controllers.ListRepos)
-	router.GET("/pull-requests/:username/:repoName/:token", controllers.PullRequests)
+	router.POST("/repositories/:token", controllers.CreateRepo)
+	router.DELETE("/repositories/:token", controllers.DeleteRepo)
+	router.GET("/users/:username/repositories", controllers.ListRepos)
+	router.GET("/repositories/:username/:repoName/pull-requests", controllers.PullRequests)
 	router.GET("/", controllers.Index)
 }
